requirements.txt
langgraph>=0.0.11
langchain-google-genai>=0.0.6
langchain-community>=0.0.13
langchain-core>=0.1.17
google-generativeai>=0.3.2
typing-extensions>=4.8.0
tavily-python>=0.3.0
python-dotenv>=0.19.0
langchain-openai>=0.0.5 